Add the option to retrieve remote licenses for projects defined in a maven pom
