+++ syft/pkg/cataloger/cataloger.go
@@ -83,7 +83,7 @@ func DirectoryCatalogers(cfg Config) []pkg.Cataloger {
 		haskell.NewHackageCataloger(),
 		java.NewArchiveCataloger(cfg.JavaConfig()),
 		java.NewGradleLockfileCataloger(),
-		java.NewPomCataloger(),
+		java.NewPomCataloger(cfg.JavaConfig()),
 		java.NewNativeImageCataloger(),
 		javascript.NewLockCataloger(cfg.Javascript),
 		nix.NewStoreCataloger(),
@@ -122,7 +122,7 @@ func AllCatalogers(cfg Config) []pkg.Cataloger {
 		haskell.NewHackageCataloger(),
 		java.NewArchiveCataloger(cfg.JavaConfig()),
 		java.NewGradleLockfileCataloger(),
-		java.NewPomCataloger(),
+		java.NewPomCataloger(cfg.JavaConfig()),
 		java.NewNativeImageCataloger(),
 		javascript.NewLockCataloger(cfg.Javascript),
 		javascript.NewPackageCataloger(),
